feat: Support word highlighting for rehype & markdown-it
