+++ packages/markdown-it-shikiji/test/fixtures/a.md
@@ -2,7 +2,7 @@
 
 …world!
 
-```js {3-4}
+```js {3-4} /a/
 console.log('it works!')
 
 const a = 1
